+++ app/tools/media/__init__.py
@@ -4,5 +4,6 @@ from .analyze_image import analyze_image
 from .generate_image import generate_image
 from .generate_video import generate_video
 from .text_to_speech import text_to_speech
+from .transcribe_audio import transcribe_audio
 
-__all__ = ["analyze_image", "generate_image", "generate_video", "text_to_speech"]
+__all__ = ["analyze_image", "generate_image", "generate_video", "text_to_speech", "transcribe_audio"]